Finished low level interrupt handling
